+++ java-generator/core/src/main/java/io/fabric8/java/generator/nodes/JCRObject.java
@@ -24,14 +24,15 @@ import com.github.javaparser.ast.expr.SingleMemberAnnotationExpr;
 import com.github.javaparser.ast.expr.StringLiteralExpr;
 import com.github.javaparser.ast.type.ClassOrInterfaceType;
 import io.fabric8.java.generator.Config;
+import io.fabric8.kubernetes.api.model.apiextensions.v1.JSONSchemaProps;
 
+import java.util.ArrayList;
 import java.util.Collections;
+import java.util.List;
+import java.util.Map;
 
-public class JCRObject extends AbstractJSONSchema2Pojo implements JObjectExtraAnnotations {
+public class JCRObject extends JObject implements JObjectExtraAnnotations {
 
-  private final String pkg;
-  private final String type;
-  private final String className;
   private final String group;
   private final String version;
   private final String scope;
@@ -53,6 +54,10 @@ public class JCRObject extends AbstractJSONSchema2Pojo implements JObjectExtraAn
       String scope,
       String specClassName,
       String statusClassName,
+      Map<String, JSONSchemaProps> toplevelProps,
+      List<String> required,
+      boolean preserveUnknownFields,
+      String description,
       boolean withSpec,
       boolean withStatus,
       boolean storage,
@@ -60,11 +65,9 @@ public class JCRObject extends AbstractJSONSchema2Pojo implements JObjectExtraAn
       String singular,
       String plural,
       Config config) {
-    super(config, null, false, null, null);
 
-    this.pkg = (pkg == null) ? "" : pkg.trim();
-    this.type = (this.pkg.isEmpty()) ? type : pkg + "." + type;
-    this.className = type;
+    super(pkg, type, toplevelProps, required, preserveUnknownFields, config, description, false, null);
+
     this.group = group;
     this.version = version;
     this.scope = scope;
@@ -146,7 +149,13 @@ public class JCRObject extends AbstractJSONSchema2Pojo implements JObjectExtraAn
     if (config.isObjectExtraAnnotations()) {
       addExtraAnnotations(clz);
     }
+
+    List<GeneratorResult.ClassResult> buffer = generateJavaFields(clz);
+
+    List<GeneratorResult.ClassResult> results = new ArrayList<>();
+    results.add(new GeneratorResult.ClassResult(className, cu));
+    results.addAll(buffer);
     return new GeneratorResult(
-        Collections.singletonList(new GeneratorResult.ClassResult(className, cu)));
+        Collections.unmodifiableList(results));
   }
 }