+++ java-generator/core/src/test/java/io/fabric8/java/generator/GeneratorTest.java
@@ -56,12 +56,16 @@ class GeneratorTest {
     // Arrange
     JCRObject cro = new JCRObject(
         "v1alpha1",
-        "t",
+        "Type",
         "g",
         "v",
         "Namespaced",
         "Spec",
         "Status",
+        Collections.emptyMap(),
+        Collections.emptyList(),
+        false,
+        "",
         true,
         true,
         true,
@@ -75,7 +79,7 @@ class GeneratorTest {
 
     // Assert
     assertEquals(1, res.getTopLevelClasses().size());
-    assertEquals("t", res.getTopLevelClasses().get(0).getName());
+    assertEquals("Type", res.getTopLevelClasses().get(0).getName());
     assertEquals("v1alpha1",
         res.getTopLevelClasses().get(0).getPackageDeclaration().get().getNameAsString());
   }
@@ -85,12 +89,16 @@ class GeneratorTest {
     // Arrange
     JCRObject cro = new JCRObject(
         null,
-        "t",
+        "Type",
         "g",
         "v",
         "Namespaced",
         "Spec",
         "Status",
+        Collections.emptyMap(),
+        Collections.emptyList(),
+        false,
+        "",
         true,
         true,
         true,
@@ -104,7 +112,7 @@ class GeneratorTest {
 
     // Assert
     assertEquals("io.fabric8.kubernetes.api.model.Namespaced", res.getTopLevelClasses().get(0)
-        .getClassByName("t").get().getImplementedTypes().get(0).getNameWithScope());
+        .getClassByName("Type").get().getImplementedTypes().get(0).getNameWithScope());
   }
 
   @Test
@@ -112,12 +120,16 @@ class GeneratorTest {
     // Arrange
     JCRObject cro = new JCRObject(
         null,
-        "t",
+        "Type",
         "g",
         "v",
         "Cluster",
         "Spec",
         "Status",
+        Collections.emptyMap(),
+        Collections.emptyList(),
+        false,
+        "",
         true,
         true,
         true,
@@ -130,7 +142,7 @@ class GeneratorTest {
     GeneratorResult res = cro.generateJava();
 
     // Assert
-    assertTrue(res.getTopLevelClasses().get(0).getClassByName("t").get().getImplementedTypes().isEmpty());
+    assertTrue(res.getTopLevelClasses().get(0).getClassByName("Type").get().getImplementedTypes().isEmpty());
   }
 
   @Test
@@ -138,12 +150,16 @@ class GeneratorTest {
     // Arrange
     JCRObject cro = new JCRObject(
         null,
-        "t",
+        "Type",
         "g",
         "v",
         "Namespaced",
         "Spec",
         "Status",
+        Collections.emptyMap(),
+        Collections.emptyList(),
+        false,
+        "",
         true,
         true,
         true,
@@ -157,7 +173,7 @@ class GeneratorTest {
 
     // Assert
     assertEquals(1, res.getTopLevelClasses().size());
-    assertEquals("t", res.getTopLevelClasses().get(0).getName());
+    assertEquals("Type", res.getTopLevelClasses().get(0).getName());
   }
 
   @Test