+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 #### Bugs
 * Fix #6038: Support for Gradle configuration cache
+* Fix #6214: Java generator does not recognize fields in CRDs other than metadata, spec, and status
 
 #### Improvements
 * Fix #5264: Remove deprecated `Config.errorMessages` field